added repo check
